# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Service API
  version: 1.0.0
  description: API contract for the inventory service MVP.
servers:
  - url: http://localhost:8080
    description: Local dev server
security:
  - SessionAuth: []
paths:
...
  /api/v1/repair-tickets:
    get:
      summary: List repair tickets
      operationId: listRepairTickets
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: equipment_id
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Repair tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RepairTicket'
    post:
      summary: Create repair ticket
      operationId: createRepairTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepairTicketCreateRequest'
      responses:
        '201':
          description: Repair ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairTicket'
        '404':
          description: Equipment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Equipment decommissioned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/repair-tickets/{id}:
    get:
      summary: Get repair ticket by id
      operationId: getRepairTicket
      parameters:
        - $ref: '#/components/parameters/RepairTicketId'
      responses:
        '200':
          description: Repair ticket found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairTicket'
        '404':
          description: Repair ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/repair-tickets/{id}/status:
    put:
      summary: Update repair ticket status
      operationId: updateRepairTicketStatus
      parameters:
        - $ref: '#/components/parameters/RepairTicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RepairTicketStatusUpdateRequest'
      responses:
        '200':
          description: Repair ticket status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepairTicket'
        '404':
          description: Repair ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Repair ticket already closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    RepairTicketId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    RepairTicket:
      type: object
      properties:
        id:
          type: integer
        equipment_id:
          type: integer
        reporter_id:
          type: integer
        issue_description:
          type: string
        urgency:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - equipment_id
        - reporter_id
        - status
    RepairTicketCreateRequest:
      type: object
      properties:
        equipment_id:
          type: integer
        issue_description:
          type: string
        urgency:
          type: string
      required:
        - equipment_id
        - issue_description
        - urgency
    RepairTicketStatusUpdateRequest:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...